+++ frontend/src/hooks/queries/useDeleteSchedule.ts
@@ -8,7 +8,7 @@ export const useDeleteSchedule = (teamPlaceId: number, scheduleId: number) => {
     () => deleteSchedule(teamPlaceId, scheduleId),
     {
       onSuccess: () => {
-        queryClient.refetchQueries({ queryKey: ['schedules'] });
+        queryClient.invalidateQueries({ queryKey: ['schedules'] });
       },
       onError: () => {
         alert('오류가 발생했습니다. 잠시 후 다시 시도해주세요.');
