+++ frontend/src/utils/formatDateTime.ts
@@ -4,5 +4,5 @@ export const formatDateTime = (rawDateTime: YYYYMMDDHHMM) => {
   const [rawDate, time] = rawDateTime.split(' ');
   const date = rawDate.split('-');
 
-  return `${date[0]}년 ${date[1]}월 ${date[0]}일 ${time}`;
+  return `${date[0]}년 ${date[1]}월 ${date[2]}일 ${time}`;
 };
